:recycle: refactor: melhorei a lógica de controller, separando arquivos
